Finalize MST project

+++ src/main/java/mst/MSTResult.java
@@ -6,9 +6,9 @@ public class MSTResult {
     public final List<Edge> mstEdges;
     public final long totalCost;
     public final long operationsCount;
-    public final double executionTimeMs;
+    public final long executionTimeMs; // long milliseconds
 
-    public MSTResult(List<Edge> mstEdges, long totalCost, long operationsCount, double executionTimeMs) {
+    public MSTResult(List<Edge> mstEdges, long totalCost, long operationsCount, long executionTimeMs) {
         this.mstEdges = mstEdges;
         this.totalCost = totalCost;
         this.operationsCount = operationsCount;
